+++ app/connectors/ReferenceDataConnector.scala
@@ -22,9 +22,7 @@ import config.FrontendAppConfig
 import connectors.ReferenceDataConnector.NoReferenceDataFoundException
 import logging.Logging
 import models.DocType.{Previous, Support, Transport}
-import models.departureTransportMeans.TransportMeansIdentification
 import models.reference._
-import models.reference.transport.TransportMode
 import models.{DocType, SecurityType}
 import play.api.http.Status._
 import play.api.libs.json.{JsError, JsResultException, JsSuccess, Reads}
