+++ app/controllers/departureMeansOfTransport/RemoveDepartureMeansOfTransportYesNoController.scala
@@ -18,8 +18,8 @@ package controllers.departureMeansOfTransport
 
 import controllers.actions._
 import forms.YesNoFormProvider
-import models.{ArrivalId, Index, Mode, TransportMeans, UserAnswers}
-import navigation.DepartureTransportMeansNavigator
+import models.removable.TransportMeans
+import models.{ArrivalId, Index, Mode, UserAnswers}
 import pages.sections.TransportMeansSection
 import play.api.data.Form
 import play.api.i18n.{I18nSupport, MessagesApi}
@@ -36,7 +36,6 @@ class RemoveDepartureMeansOfTransportYesNoController @Inject() (
   sessionRepository: SessionRepository,
   actions: Actions,
   formProvider: YesNoFormProvider,
-  navigator: DepartureTransportMeansNavigator,
   val controllerComponents: MessagesControllerComponents,
   view: RemoveDepartureMeansOfTransportYesNoView
 )(implicit ec: ExecutionContext)
