+++ app/views/departureMeansOfTransport/IdentificationView.scala.html
@@ -15,7 +15,7 @@
  *@
 
 @import models.MovementReferenceNumber
-@import models.departureTransportMeans.TransportMeansIdentification
+@import models.reference.TransportMeansIdentification
 @import viewModels.components.InputRadioViewModel._
 @import views.html.components.{Button, ErrorSummary, InputRadio}
 @import viewModels.departureTransportMeans.IdentificationViewModel
