+++ test/models/TransportMeansSpec.scala
@@ -18,7 +18,8 @@ package models
 
 import base.SpecBase
 import generators.Generators
-import models.departureTransportMeans.TransportMeansIdentification
+import models.reference.TransportMeansIdentification
+import models.removable.TransportMeans
 import org.scalacheck.Arbitrary.arbitrary
 import org.scalacheck.Gen
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
