+++ test/viewModels/departureTransportMeans/AddAnotherDepartureMeansOfTransportViewModelSpec.scala
@@ -18,8 +18,7 @@ package viewModels.departureTransportMeans
 
 import base.SpecBase
 import generators.Generators
-import models.departureTransportMeans.TransportMeansIdentification
-import models.reference.Country
+import models.reference.{Country, TransportMeansIdentification}
 import models.{Index, Mode}
 import org.scalacheck.Arbitrary.arbitrary
 import org.scalacheck.Gen
